Refactor package location for anitrend.arch extensions
