+++ app-core/src/main/kotlin/co/anitrend/core/glide/GlideAppModule.kt
@@ -28,7 +28,7 @@ import com.bumptech.glide.load.engine.cache.LruResourceCache
 import com.bumptech.glide.load.engine.cache.MemorySizeCalculator
 import com.bumptech.glide.module.AppGlideModule
 import com.bumptech.glide.request.RequestOptions
-import co.anitrend.arch.extension.isLowRamDevice
+import co.anitrend.arch.extension.ext.isLowRamDevice
 
 @GlideModule
 class GlideAppModule: AppGlideModule() {
