+++ app-core/src/main/kotlin/co/anitrend/core/ui/fragment/list/AniTrendListFragment.kt
@@ -18,8 +18,8 @@
 package co.anitrend.core.ui.fragment.list
 
 import android.content.Context
-import co.anitrend.arch.extension.attachComponent
-import co.anitrend.arch.extension.detachComponent
+import co.anitrend.arch.extension.ext.attachComponent
+import co.anitrend.arch.extension.ext.detachComponent
 import co.anitrend.arch.ui.fragment.list.SupportFragmentList
 import co.anitrend.core.ui.ILifecycleFeature
