+++ app-core/src/main/kotlin/co/anitrend/core/util/locale/AniTrendLocale.kt
@@ -17,7 +17,7 @@
 
 package co.anitrend.core.util.locale
 
-import co.anitrend.arch.extension.empty
+import co.anitrend.arch.extension.ext.empty
 import java.util.Locale
 
 /**
