+++ app-data/src/main/kotlin/co/anitrend/data/tag/source/MediaTagSourceImpl.kt
@@ -17,7 +17,7 @@
 
 package co.anitrend.data.tag.source
 
-import co.anitrend.arch.extension.SupportDispatchers
+import co.anitrend.arch.extension.dispatchers.SupportDispatchers
 import co.anitrend.data.arch.controller.strategy.policy.OnlineStrategy
 import co.anitrend.data.arch.extension.controller
 import co.anitrend.data.arch.helper.data.ClearDataHelper
@@ -26,7 +26,6 @@ import co.anitrend.data.tag.datasource.local.MediaTagLocalSource
 import co.anitrend.data.tag.datasource.remote.MediaTagRemoteSource
 import co.anitrend.data.tag.mapper.MediaTagResponseMapper
 import co.anitrend.data.tag.source.contract.MediaTagSource
-import kotlinx.coroutines.ExperimentalCoroutinesApi
 import kotlinx.coroutines.async
 import kotlinx.coroutines.flow.flowOn
 import kotlinx.coroutines.flow.map
@@ -40,7 +39,6 @@ internal class MediaTagSourceImpl(
     dispatchers: SupportDispatchers
 ) : MediaTagSource(dispatchers) {
 
-    @ExperimentalCoroutinesApi
     override val observable =
         localSource.findAllFlow().map {
             converter.convertFrom(it)
