+++ app-data/src/main/kotlin/co/anitrend/data/tag/source/contract/MediaTagSource.kt
@@ -21,7 +21,7 @@ import androidx.lifecycle.LiveData
 import androidx.lifecycle.asLiveData
 import androidx.lifecycle.liveData
 import co.anitrend.arch.data.source.coroutine.SupportCoroutineDataSource
-import co.anitrend.arch.extension.SupportDispatchers
+import co.anitrend.arch.extension.dispatchers.SupportDispatchers
 import co.anitrend.domain.tag.entities.Tag
 import kotlinx.coroutines.flow.Flow
