+++ common-media-ui/src/main/kotlin/co/anitrend/common/media/ui/widgets/MediaStatusWidget.kt
@@ -20,7 +20,7 @@ package co.anitrend.common.media.ui.widgets
 import android.content.Context
 import android.util.AttributeSet
 import android.widget.FrameLayout
-import co.anitrend.arch.extension.getCompatDrawable
+import co.anitrend.arch.extension.ext.getCompatDrawable
 import co.anitrend.arch.ui.view.contract.CustomView
 import co.anitrend.common.media.ui.R
 import co.anitrend.domain.media.enums.MediaStatus
